refactor: arrange backend function files
